feb9save

+++ lib/data/model/auth/get_all/ftx_get_balance_model.dart
@@ -28,7 +28,7 @@ class Result {
   double availableWithoutBorrow;
   String coin;
   double free;
-  int spotBorrow;
+  double spotBorrow;
   double total;
   double usdValue;
